LayoutTests/js/array-storage-splice-holes-require-special-behavior-expected.txt

Test to ensure correct behaviour of splice in array storage mode with indexed properties in the prototype chain.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a[0] is 1
PASS a[1] is 2
PASS a[2] is 3
PASS a[3] is 5
PASS a[4] is 6
PASS a[5] is 7
PASS a[6] is 8
PASS a[7] is 9
PASS b[0] is 1
PASS b[1] is 2
PASS b[2] is 3
PASS b[3] is 5
PASS b[4] is 6
PASS b[5] is 7
PASS b[6] is 8
PASS b[7] is 9
PASS successfullyParsed is true

TEST COMPLETE
